+++ web/wasm24/wasm24.d.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 /**
 * @param {number} target
-* @param {Uint8Array} list
+* @param {Uint16Array} list
 * @returns {string | undefined}
 */
-export function solve(target: number, list: Uint8Array): string | undefined;
+export function solve(target: number, list: Uint16Array): string | undefined;
